services/service2/requirements.txt
flask
prometheus_client
requests
python-telegram-bot
